+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.3.1
+* Fix issues completion is case sensitive [#32](https://github.com/zabel-xyz/plsql-language/issues/37)
+* Fix issues completion - different icon for symbols [#32](https://github.com/zabel-xyz/plsql-language/issues/37)
+* Fix issues completion with loading hangs [#32](https://github.com/zabel-xyz/plsql-language/issues/37)
+
 ## 1.3.0
 * Fix issues with regExp parser (forward declaration and body declaration) [#37](https://github.com/zabel-xyz/plsql-language/issues/37)
 * Add intelliSense for package members (autocompletion from package files) [#32](https://github.com/zabel-xyz/plsql-language/issues/32)